Fix - Move candidature creation BL to controller
