+++ src/controllers/candidatures.ts
@@ -1,6 +1,8 @@
 import Candidature from '../models/candidature';
 import PastYearExp from '../models/pastyearexp';
 import Attachment from '../models/attachment';
+import User from '../models/user';
+import logger from '../helpers/logger';
 
 async function getAll(): Promise<Candidature[]> {
   return Candidature.findAll({
@@ -54,4 +56,51 @@ function getById(id: number, role: string): Promise<Candidature> {
   return Candidature.findByPk(id, options);
 }
 
-export = { getAll, getById }
+async function createCandidature(user: User, params: any): Promise<Candidature | checkError[]> {
+  const [valid, errs] = isValid(params);
+  if (valid) {
+    const attachments = params.attachments;
+    const experiences = params.experiences;
+    delete params.attachments;
+    delete params.experiences;
+    return user!.createCandidature(params)
+      .then(async cand => {
+        //We need to create each attachment and experiences
+        const promises: Promise<Attachment | PastYearExp>[] = [];
+
+        if (attachments !== undefined) promises.push(attachments.map((attach: any) => cand.createAttachment(attach)));
+        if (experiences !== undefined) promises.push(experiences.map((exp: any) => cand.createExperience(exp)));
+
+        await Promise.all(promises);
+
+        return cand;
+      })
+      .catch(err => {
+        logger.error(['Error while creating an application', err]);
+        throw err;
+      });
+  } else {
+    return errs;
+  }
+}
+
+
+interface checkError {
+  id: string,
+  error: string
+}
+
+function isValid(reqBody: any): [boolean, checkError[]] {
+  if (<boolean>reqBody.draft) {
+    return [true, []];
+  } else {
+    let errs: checkError[] = [];
+
+    // Check if all information are correct TODO
+    errs.push({ id: 'first_name', error: 'Une erreur' });
+
+    return errs.length === 0 ? [true, []] : [false, errs];
+  }
+}
+
+export = { getAll, getById, createCandidature }
