+++ src/routes/candidature.ts
@@ -4,8 +4,7 @@ import Candidature from '../models/candidature';
 import User from '../models/user';
 import checkJwt from '../middlewares/auth.middleware';
 import logger from '../helpers/logger';
-import Attachment from '../models/attachment';
-import PastYearExp from '../models/pastyearexp';
+
 const candidatureRouter = Router();
 
 candidatureRouter.get('/', [checkJwt], async (req: Request, res: Response) => {
@@ -57,57 +56,27 @@ candidatureRouter.post('/', [checkJwt], async (req: Request, res: Response) => {
   const user = await User.findByPk(userId);
   if (user === undefined) {
     logger.error(`User ${userId} not found while trying to create an application`);
-    res.status(404).send('Utilisateur non trouvé');
+    res
+      .status(404)
+      .send('Utilisateur non trouvé');
   } else {
     //role guard
     if (user!.role !== 'eleve') return res.status(403).send('Vous ne pouvez pas créer une candidature sans être élève');
 
-    const [valid, errs] = isValid(params);
-    if (valid) {
-      const attachments = params.attachments;
-      const experiences = params.experiences;
-      delete params.attachments;
-      delete params.experiences;
-      user!.createCandidature(params)
-        .then(async cand => {
-          //We need to create each attachment and experiences
-          const promises: Promise<Attachment | PastYearExp>[] = [];
-
-          if (attachments !== undefined) promises.push(attachments.map((attach: any) => cand.createAttachment(attach)));
-          if (experiences !== undefined) promises.push(experiences.map((exp: any) => cand.createExperience(exp)));
-
-          await Promise.all(promises);
-
-          res.status(201).json(cand);
-        })
-        .catch(err => {
-          logger.error(['Error while creating an application', err]);
-          res.status(500).send(err.message);
-        });
-    } else {
-      res
-        .status(400)
-        .json({ errors: errs });
+    try {
+      const creationResponse = await candidatureController.createCandidature(user!, params);
+      if (creationResponse instanceof Candidature) {
+        res.sendStatus(201);
+      } else {
+        res
+          .status(400)
+          .json({ errors: creationResponse });
+      }
+    } catch (err) {
+      res.status(500).send(err.message);
     }
   }
 });
 
-interface checkError {
-  id: string,
-  error: string
-}
-
-function isValid(reqBody: any): [boolean, checkError[]] {
-  if (<boolean>reqBody.draft) {
-    return [true, []];
-  } else {
-    let errs: checkError[] = [];
-
-    // Check if all information are correct TODO
-    errs.push({ id: 'first_name', error: 'Une erreur' });
-
-    return errs.length === 0 ? [true, []] : [false, errs];
-  }
-}
 
 export = candidatureRouter;
